removed dashCasing from yoast for graphql
